float to double

+++ R2+1D/AdaptiveAvgPool3d.cpp
@@ -1,5 +1,5 @@
 #include "r2plus1d.h"
-void AdaptiveAvgPool3d(float* X_data,int* X_num, int* out_size, float* Y_data, int* Y_num ){
+void AdaptiveAvgPool3d(double* X_data,int* X_num, int* out_size, double* Y_data, int* Y_num ){
     int N = X_num[0];
 	int C = X_num[1];
 	int D = X_num[2];
