+++ R2+1D/r2plus1d.cpp
@@ -2,11 +2,11 @@
 #include<iostream>
 using namespace std;
 
-void r2plus1d(float* X_data, float* Y_data, float* Kernel_1_data, float* Kernel_2_data)
+void r2plus1d(double* X_data, double* Y_data, double* Kernel_1_data, double* Kernel_2_data)
 {
  #ifdef __SYNTHESIS__
-    float X_out_data[141120];
-    float Y_tmp_data[200704];
+    double X_out_data[141120];
+    double Y_tmp_data[200704];
 
     #pragma HLS INTERFACE s_axilite port=return
 	#pragma HLS INTERFACE m_axi port=X_data
@@ -14,8 +14,8 @@ void r2plus1d(float* X_data, float* Y_data, float* Kernel_1_data, float* Kernel_
     #pragma HLS INTERFACE m_axi port=Kernel_1_data
 	#pragma HLS INTERFACE m_axi port=Kernel_2_data
  #else
-     float* X_out_data = (float*)malloc(141120*sizeof(float)); // value after first Conv-Batch-ReLU layer
-     float* Y_tmp_data = (float*)malloc(200704*sizeof(float)); // value after second Conv-Batch-ReLU layer
+     double* X_out_data = (double*)malloc(141120*sizeof(double)); // value after first Conv-Batch-ReLU layer
+     double* Y_tmp_data = (double*)malloc(200704*sizeof(double)); // value after second Conv-Batch-ReLU layer
  #endif
 
     // R2Plus1dStem
